Prism/backend/pyproject.toml
[project]
name = "prism-backend"
version = "0.1.0"
description = "Prism Document Q&A Backend API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",
    "aiofiles==23.2.1",
    "python-dotenv==1.0.0",
    "PyPDF2==3.0.1",
    "python-docx==1.1.0",
    "tiktoken==0.5.1",
    "langchain-text-splitters==0.0.1",
    "pydantic==2.10.3",
    "pydantic-core==2.27.1",
    "requests==2.32.3",
    "anyio==3.7.1",
    "watchfiles==1.1.1",
    "starlette==0.27.0",
    "httptools==0.7.1",
    "websockets==15.0.1",
    "orjson==3.11.3",
    "tenacity==8.5.0",
    "python-json-logger==2.0.7",
    "chromadb",
    "sentence-transformers",
    "paddlepaddle",
    "paddleocr>=2.7.0",
    "opencv-python-headless"
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.0",
    "pytest-asyncio==0.24.0",
    "pytest-cov==4.1.0",
    "httpx==0.28.1",
    "black==24.0.0",
    "isort==5.13.0",
    "flake8==7.0.0",
    "mypy==1.8.0",
    "pre-commit==3.6.0"
]

test = [
    "pytest==8.0.0",
    "pytest-asyncio==0.24.0",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "httpx==0.28.1",
    "coverage[toml]==7.4.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
